+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/Entities.hs
@@ -18,6 +18,7 @@ module Cardano.Ledger.Dijkstra.Rules.Entities (
   EntitiesEnv (..),
   EntitiesPredFailure (..),
   EntitiesEvent (..),
+  validateWrongNetworkInDirectDeposit,
 ) where
 
 import Cardano.Ledger.Address (DirectDeposits (..))
@@ -32,6 +33,7 @@ import Cardano.Ledger.Dijkstra.Era (DijkstraEra, ENTITIES)
 import Cardano.Ledger.Dijkstra.Rules.Certs ()
 import Cardano.Ledger.Dijkstra.Rules.GovCert (DijkstraGovCertPredFailure)
 import Cardano.Ledger.Dijkstra.TxBody (DijkstraEraTxBody, directDepositsTxBodyL)
+import Cardano.Ledger.Rules.ValidationMode (Test, runTest)
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
 import Control.DeepSeq (NFData)
 import Control.Monad.Trans.Reader (asks)
@@ -40,6 +42,7 @@ import Data.Map.NonEmpty (NonEmptyMap)
 import qualified Data.Map.NonEmpty as NEM
 import qualified Data.Map.Strict as Map
 import Data.Sequence (Seq)
+import Data.Set.NonEmpty (NonEmptySet)
 import GHC.Generics (Generic)
 import Lens.Micro
 
@@ -65,10 +68,20 @@ instance EraTx era => EncCBOR (EntitiesEnv era) where
 
 data EntitiesPredFailure era
   = CertsFailure (PredicateFailure (EraRule "CERTS" era))
-  | WithdrawalsMissingAccounts Withdrawals
+  | MissingAccountsInWithdrawals Withdrawals
   | IncompleteWithdrawals (NonEmptyMap AccountAddress (Mismatch RelEQ Coin))
-  | WithdrawalAmountsExceedAccountBalances (NonEmptyMap AccountAddress (Mismatch RelLTEQ Coin))
-  | DirectDepositsToMissingAccounts DirectDeposits
+  | ExceededBalancesInWithdrawals (NonEmptyMap AccountAddress (Mismatch RelLTEQ Coin))
+  | MissingAccountsInDirectDeposits DirectDeposits
+  | WrongNetworkInWithdrawals
+      -- | Expected network id
+      Network
+      -- | Withdrawal accounts with wrong network id
+      (NonEmptySet AccountAddress)
+  | WrongNetworkInDirectDeposits
+      -- | Expected network id
+      Network
+      -- | Direct-deposit accounts with wrong network id
+      (NonEmptySet AccountAddress)
   deriving (Generic)
 
 deriving stock instance
@@ -90,10 +103,12 @@ instance
   encCBOR =
     encode . \case
       CertsFailure x -> Sum (CertsFailure @era) 0 !> To x
-      WithdrawalsMissingAccounts x -> Sum (WithdrawalsMissingAccounts @era) 1 !> To x
+      MissingAccountsInWithdrawals x -> Sum (MissingAccountsInWithdrawals @era) 1 !> To x
       IncompleteWithdrawals x -> Sum (IncompleteWithdrawals @era) 2 !> To x
-      WithdrawalAmountsExceedAccountBalances x -> Sum (WithdrawalAmountsExceedAccountBalances @era) 3 !> To x
-      DirectDepositsToMissingAccounts x -> Sum (DirectDepositsToMissingAccounts @era) 4 !> To x
+      ExceededBalancesInWithdrawals x -> Sum (ExceededBalancesInWithdrawals @era) 3 !> To x
+      MissingAccountsInDirectDeposits x -> Sum (MissingAccountsInDirectDeposits @era) 4 !> To x
+      WrongNetworkInWithdrawals x y -> Sum (WrongNetworkInWithdrawals @era) 5 !> To x !> To y
+      WrongNetworkInDirectDeposits x y -> Sum (WrongNetworkInDirectDeposits @era) 6 !> To x !> To y
 
 instance
   ( Era era
@@ -103,10 +118,12 @@ instance
   where
   decCBOR = decode . Summands "EntitiesPredFailure" $ \case
     0 -> SumD CertsFailure <! From
-    1 -> SumD WithdrawalsMissingAccounts <! From
+    1 -> SumD MissingAccountsInWithdrawals <! From
     2 -> SumD IncompleteWithdrawals <! From
-    3 -> SumD WithdrawalAmountsExceedAccountBalances <! From
-    4 -> SumD DirectDepositsToMissingAccounts <! From
+    3 -> SumD ExceededBalancesInWithdrawals <! From
+    4 -> SumD MissingAccountsInDirectDeposits <! From
+    5 -> SumD WrongNetworkInWithdrawals <! From <! From
+    6 -> SumD WrongNetworkInDirectDeposits <! From <! From
     n -> Invalid n
 
 newtype EntitiesEvent era = CertsEvent (Event (EraRule "CERTS" era))
@@ -143,6 +160,9 @@ instance InjectRuleFailure "ENTITIES" DijkstraGovCertPredFailure DijkstraEra whe
 instance InjectRuleFailure "ENTITIES" Conway.ConwayLedgerPredFailure DijkstraEra where
   injectFailure = conwayToDijkstraEntitiesPredFailure
 
+instance InjectRuleFailure "ENTITIES" Shelley.ShelleyUtxoPredFailure DijkstraEra where
+  injectFailure = shelleyUtxoToDijkstraEntitiesPredFailure
+
 instance
   ( EraTx era
   , DijkstraEraTxBody era
@@ -154,6 +174,7 @@ instance
   , Environment (EraRule "CERTS" era) ~ Conway.CertsEnv era
   , EraRule "ENTITIES" era ~ ENTITIES era
   , InjectRuleFailure "ENTITIES" EntitiesPredFailure era
+  , InjectRuleFailure "ENTITIES" Shelley.ShelleyUtxoPredFailure era
   , InjectRuleFailure "ENTITIES" Conway.ConwayLedgerPredFailure era
   ) =>
   STS (ENTITIES era)
@@ -179,6 +200,7 @@ dijkstraEntitiesTransition ::
   , Environment (EraRule "CERTS" era) ~ Conway.CertsEnv era
   , EraRule "ENTITIES" era ~ ENTITIES era
   , InjectRuleFailure "ENTITIES" EntitiesPredFailure era
+  , InjectRuleFailure "ENTITIES" Shelley.ShelleyUtxoPredFailure era
   , InjectRuleFailure "ENTITIES" Conway.ConwayLedgerPredFailure era
   ) =>
   TransitionRule (ENTITIES era)
@@ -189,6 +211,9 @@ dijkstraEntitiesTransition = do
       accounts = certState ^. certDStateL . accountsL
 
   network <- liftSTS $ asks networkId
+
+  runTest $ Shelley.validateWrongNetworkWithdrawal network (tx ^. bodyTxL)
+  runTest $ validateWrongNetworkInDirectDeposit network (tx ^. bodyTxL)
 
   validateWithdrawals legacyMode network withdrawals accounts
 
@@ -203,9 +228,23 @@ dijkstraEntitiesTransition = do
   let directDeposits = tx ^. bodyTxL . directDepositsTxBodyL
       accountsAfterCerts = certStateAfterCerts ^. certDStateL . accountsL
   failOnJust (directDepositsMissingAccounts directDeposits accountsAfterCerts) $
-    injectFailure . DirectDepositsToMissingAccounts
+    injectFailure . MissingAccountsInDirectDeposits
 
   pure $ certStateAfterCerts & certDStateL . accountsL %~ applyDirectDeposits directDeposits
+
+validateWrongNetworkInDirectDeposit ::
+  DijkstraEraTxBody era =>
+  Network ->
+  TxBody t era ->
+  Test (EntitiesPredFailure era)
+validateWrongNetworkInDirectDeposit netId txb =
+  failureOnNonEmptySet depositsWrongNetwork (WrongNetworkInDirectDeposits netId)
+  where
+    depositsWrongNetwork =
+      Map.keysSet $
+        Map.filterWithKey
+          (\a _ -> aaNetworkId a /= netId)
+          (unDirectDeposits $ txb ^. directDepositsTxBodyL)
 
 validateWithdrawals ::
   EraAccounts era =>
@@ -229,10 +268,10 @@ validateWithdrawals legacyMode network withdrawals accounts = do
               case withdrawalsThatExceedAccountBalance withdrawals network accounts of
                 Nothing -> (Map.empty, Map.empty)
                 Just (missing, exceeded) -> (unWithdrawals missing, exceeded)
-        failOnNonEmptyMap exceededWithdrawals WithdrawalAmountsExceedAccountBalances
+        failOnNonEmptyMap exceededWithdrawals ExceededBalancesInWithdrawals
         pure missingWithdrawals
   failOnNonEmptyMap missingWithdrawals $
-    WithdrawalsMissingAccounts . Withdrawals . NEM.toMap
+    MissingAccountsInWithdrawals . Withdrawals . NEM.toMap
 
 conwayToDijkstraEntitiesPredFailure ::
   forall era. Conway.ConwayLedgerPredFailure era -> EntitiesPredFailure era
@@ -246,6 +285,23 @@ conwayToDijkstraEntitiesPredFailure = \case
   Conway.ConwayMempoolFailure _ -> impossible "ConwayMempoolFailure"
   Conway.ConwayWithdrawalsMissingAccounts _ -> impossible "ConwayWithdrawalsMissingAccounts"
   Conway.ConwayIncompleteWithdrawals _ -> impossible "ConwayIncompleteWithdrawals"
+  where
+    impossible name = error $ "Impossible: `" <> name <> "` for ENTITIES"
+
+shelleyUtxoToDijkstraEntitiesPredFailure ::
+  Shelley.ShelleyUtxoPredFailure era -> EntitiesPredFailure era
+shelleyUtxoToDijkstraEntitiesPredFailure = \case
+  Shelley.WrongNetworkWithdrawal net addrs -> WrongNetworkInWithdrawals net addrs
+  Shelley.BadInputsUTxO _ -> impossible "BadInputsUTxO"
+  Shelley.ExpiredUTxO _ -> impossible "ExpiredUTxO"
+  Shelley.MaxTxSizeUTxO _ -> impossible "MaxTxSizeUTxO"
+  Shelley.InputSetEmptyUTxO -> impossible "InputSetEmptyUTxO"
+  Shelley.FeeTooSmallUTxO _ -> impossible "FeeTooSmallUTxO"
+  Shelley.ValueNotConservedUTxO _ -> impossible "ValueNotConservedUTxO"
+  Shelley.WrongNetwork _ _ -> impossible "WrongNetwork"
+  Shelley.OutputTooSmallUTxO _ -> impossible "OutputTooSmallUTxO"
+  Shelley.UpdateFailure _ -> impossible "UpdateFailure"
+  Shelley.OutputBootAddrAttrsTooBig _ -> impossible "OutputBootAddrAttrsTooBig"
   where
     impossible name = error $ "Impossible: `" <> name <> "` for ENTITIES"