+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Imp/EntitiesSpec.hs
@@ -6,9 +6,11 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 
-module Test.Cardano.Ledger.Dijkstra.Imp.CertsSpec (spec) where
+module Test.Cardano.Ledger.Dijkstra.Imp.EntitiesSpec (spec) where
 
-import Cardano.Ledger.BaseTypes (EpochInterval (..), Mismatch (..))
+import Cardano.Base.Typeable (Typeable)
+import Cardano.Ledger.Address
+import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.DRep (DRep (..))
@@ -16,17 +18,21 @@ import Cardano.Ledger.Dijkstra.Core
 import Cardano.Ledger.Dijkstra.Rules (
   DijkstraUtxoPredFailure (..),
   EntitiesPredFailure (..),
+  SubEntitiesPredFailure (..),
  )
 import Cardano.Ledger.Plutus
 import Cardano.Ledger.Val (Val (..))
+import qualified Data.Foldable as Foldable
+import Data.List ((\\))
 import qualified Data.Map.NonEmpty as NE
+import qualified Data.Set.NonEmpty as NES
 import Lens.Micro ((&), (.~))
 import Test.Cardano.Ledger.Dijkstra.ImpTest
 import Test.Cardano.Ledger.Imp.Common
 import Test.Cardano.Ledger.Plutus.Examples (alwaysSucceedsWithDatum)
 
 spec :: forall era. DijkstraEraImp era => SpecWith (ImpInit (LedgerSpec era))
-spec = describe "CERTS" $ do
+spec = describe "ENTITIES" $ do
   describe "Withdrawals" $ do
     it "Withdrawing from an unregistered staking address" $ do
       modifyPParams $ ppGovActionLifetimeL .~ EpochInterval 2
@@ -45,7 +51,7 @@ spec = describe "CERTS" $ do
             WithdrawalsExceedAccountBalance @era $
               NE.singleton accountAddress $
                 Mismatch (Coin 20) mempty
-        , injectFailure . WithdrawalsMissingAccounts @era $
+        , injectFailure . MissingAccountsInWithdrawals @era $
             Withdrawals [(accountAddress, Coin 20)]
         ]
       (registeredAccountAddress, reward, stakeKey2) <- setupAccountAddress
@@ -58,7 +64,7 @@ spec = describe "CERTS" $ do
                 .~ Withdrawals [(accountAddress, zero), (registeredAccountAddress, reward)]
       submitFailingTx
         tx2
-        [ injectFailure . WithdrawalsMissingAccounts @era $
+        [ injectFailure . MissingAccountsInWithdrawals @era $
             Withdrawals [(accountAddress, zero)]
         ]
 
@@ -83,7 +89,7 @@ spec = describe "CERTS" $ do
               NE.singleton accountAddress1 $
                 Mismatch (reward1 <+> Coin 1) reward1
         , injectFailure $
-            WithdrawalAmountsExceedAccountBalances @era $
+            ExceededBalancesInWithdrawals @era $
               NE.singleton accountAddress1 $
                 Mismatch (reward1 <+> Coin 1) reward1
         ]
@@ -109,6 +115,32 @@ spec = describe "CERTS" $ do
             & withdrawalsTxBodyL
               .~ Withdrawals
                 [(accountAddress1, zero)]
+
+    it "Rejects withdrawals and direct deposits with wrong network id" $ do
+      stakeKey <- freshKeyHash
+      accountAddress <- registerStakeCredential (KeyHashObj stakeKey)
+      let wrongNetworkAccount = accountAddress & accountAddressNetworkIdL .~ Mainnet
+      let txBody :: forall l. Typeable l => TxBody l era
+          txBody =
+            mkBasicTxBody
+              & withdrawalsTxBodyL
+                .~ Withdrawals [(wrongNetworkAccount, mempty)]
+              & directDepositsTxBodyL
+                .~ DirectDeposits [(wrongNetworkAccount, Coin 50)]
+
+      submitFailingTx
+        (mkBasicTx txBody)
+        [ injectFailure . WrongNetworkInWithdrawals @era Testnet $ NES.singleton wrongNetworkAccount
+        , injectFailure . WrongNetworkInDirectDeposits @era Testnet $ NES.singleton wrongNetworkAccount
+        , injectFailure . MissingAccountsInWithdrawals @era $ Withdrawals [(wrongNetworkAccount, mempty)]
+        ]
+
+      submitFailingTxIncluding
+        (mkBasicTx $ txBody & subTransactionsTxBodyL .~ [mkBasicTx txBody])
+        [ injectFailure . SubWrongNetworkInWithdrawals @era Testnet $ NES.singleton wrongNetworkAccount
+        , injectFailure . SubWrongNetworkInDirectDeposits @era Testnet $ NES.singleton wrongNetworkAccount
+        , injectFailure . SubMissingAccountsInWithdrawals @era $ Withdrawals [(wrongNetworkAccount, mempty)]
+        ]
   where
     setupAccountAddress :: ImpTestM era (AccountAddress, Coin, KeyHash Staking)
     setupAccountAddress = do
@@ -118,3 +150,13 @@ spec = describe "CERTS" $ do
       submitAndExpireProposalToMakeReward cred
       b <- getBalance cred
       pure (ra, b, kh)
+
+    -- Poor man's `submitFailingTx` - only checking that the given predicate failures
+    -- are part of all the failures returned by submitting the transaction.
+    -- TOOD: to be replaced by `submitFailingTx` when the Imp fixup for nested transaction is done.
+    submitFailingTxIncluding tx expected = do
+      result <- trySubmitTx tx
+      case result of
+        Left (predFailures, _) ->
+          (expected \\ Foldable.toList predFailures) `shouldBe` []
+        Right _ -> expectationFailure "Expected submission to fail, but it succeeded"